\name{IAM.model}
\alias{IAM.model}
\alias{IAM.model-methods}
\alias{IAM.model,iamArgs,iamInput-method}
\docType{methods}

\title{Modelization method}
\description{ToDo}
\section{Methods}{
  \describe{
	\item{IAM.model}{\code{signature(objArgs = "iamArgs", objInput = "iamInput"}: objArgs & objInput are outputs from, respectively, \emph{IAM.args} and \emph{IAM.input} methods.}
}}


\keyword{methods}


