\name{iamArgs-class}
\docType{class}
\alias{iamArgs-class}
\title{Class "iamArgs"}
\description{ToDo}

\section{Slots}{
\describe{
    \item{\code{desc}:}{object description}
    \item{\code{arguments}:}{ToDo}
    \item{\code{specific}:}{ToDo}
  }
}

\author{Mathieu Merzereaud}
\examples{
showClass("iamArgs")
}
\keyword{classes}

