\name{iamData}
\alias{iamData}
\alias{DFiam}
\alias{DFiam_iter}
\docType{data}
\title{Formatted IAM output illustrating plotting functions}
\description{
Typical IAM.unite output object to illustrate IAM plotting functions.   
}
\usage{DFiam}
%\usage{DFiam_iter}
\format{Data.frame with replicated example value at time, scenario, and age (as headers)}
\section{WARNING}{
This data is provided is for testing only and should not be used for any other purpose. 
}
\details{
To Do
}
\source{
To Do
}
\examples{
data(iamData)
}
\keyword{datasets}

