\name{iamInput-class}
\docType{class}
\alias{iamInput-class}
\title{Class "iamInput"}
\description{ToDo}

\section{Slots}{
\describe{
    \item{\code{desc}:}{object description}
    \item{\code{specific}:}{ToDo}
    \item{\code{historical}:}{ToDo}
    \item{\code{input}:}{ToDo}
    \item{\code{scenario}:}{ToDo}
    \item{\code{stochastic}:}{ToDo}
    \item{\code{optimization}:}{ToDo}
  }
}

\author{Mathieu Merzereaud}
\examples{
showClass("iamInput")
}
\keyword{classes}

