\name{iamOutputRep-class}
\docType{class}
\alias{iamOutputRep-class}
\title{Class "iamOutputRep"}
\description{ToDo}

\section{Slots}{
\describe{
    \item{\code{desc}:}{object description}
    \item{\code{arguments}:}{ToDo}
    \item{\code{specific}:}{ToDo}
    \item{\code{outputSp}:}{ToDo}
    \item{\code{output}:}{ToDo}
  }
}

\author{Mathieu Merzereaud}
\examples{
showClass("iamOutputRep")
}
\keyword{classes}

